minidesignfix: escaping in evalable terms (2/n)
The prior design (undocumented) used backslash (\) to escape dollar signs ($) so that ${HOME}, etc. can be escaped. But that means Windows paths will contain escapes.

Now the lexing rules are documented, and a different escape character chosen.
We could have used ^ (Windows Batch) or ` (PowerShell) for familiarity and compatibility with Windows. We chose the backtick ` for compatibility with Windows paths and familiarity with PowerShell.